Howto write a use case diagram:
First, create the diagram from the top of your head.
Then delete the file. Since you used ext3, you don't have a chance to recover it.
But you notice that only hours later.
You start again on a sheet (which seems safer).
You create the diagram again, noticing it is more complete and of better quality than the first one.


git-svn-id: svn://svn.jakeapp.com/trunk@14 cfed719c-3465-4289-a5de-482d64f03f8e